requirements.txt
google-api-python-client==2.95.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.4
python-dotenv==1.0.0
openai==1.3.0
pandas==2.0.3
schedule==1.2.0
